refactor: Move interactive commands to interactive.rs
